Remove precompile logic from barbican
This change removes the necessity to ran a make task before a chart can be packaged.

We should just use {{.Files.Get}} for everything that is not a template (patches configmap and probaply the bin cm as well).

For the etc config map I moved the files into the templates folder and referenced them acordingly.