[nova] enable_new_services option
Give all services a fixed hostname and enable them on install, but not on uprade.
This should ensure, that new services are taken online after a validation, which is outside of the scope of helm.
(E.g compute resources)
This can be overridden with the corresponding value enable_new_services=True